Renamed components for clarity
